requirements.txt
asgiref==3.11.0
Django==6.0
dnspython==2.8.0
pymongo==4.15.5
python-dotenv==1.2.1
sqlparse==0.5.4
tzdata==2025.2
